requirements.txt
uvicorn~=0.29.0
fastapi~=0.111.0
opencv-python~=4.9.0.80
pycocotools~=2.0.7
torch~=2.3.0
torchvision~=0.18.0
PyYAML~=6.0.1
tqdm~=4.66.4
pandas~=2.2.2
seaborn~=0.13.2
requests~=2.31.0
gitpython
ipython
matplotlib>=3.2.2
numpy>=1.18.5
Pillow>=7.1.2
psutil
scipy>=1.4.1
thop>=0.1.1
tensorboard>=2.4.1
albumentations>=1.0.3
git+https://github.com/WhyNotHugo/python-barcode
timm==0.6.11
wandb
torch-summary-

